Day35/LoginApp/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib
python-jose[cryptography]
python-dotenv
email-validator
python-multipart